+++ apps/src/pythonlab/layout/ConsoleAndPreviewShare.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 
 import MiniAppPreview from '@cdo/apps/codebridge/MiniAppPreview/MiniAppPreview';
 import HorizontalOutput from '@cdo/apps/codebridge/Workspace/HorizontalOutput';
@@ -7,17 +7,35 @@ import moduleStyles from './share-view.module.scss';
 
 interface ConsoleAndPreviewShareProps {
   consoleVisible: boolean;
-  availableWidth: number;
 }
 
 const ConsoleAndPreviewShare: React.FunctionComponent<
   ConsoleAndPreviewShareProps
-> = ({consoleVisible, availableWidth}) => {
+> = ({consoleVisible}) => {
+  const getAvailableOutputWidth = () => {
+    return Math.max(window.innerWidth - 150, 400);
+  };
+
+  const [outputWidth, setOutputWidth] = useState(getAvailableOutputWidth());
+  const [outputHeight, setOutputHeight] = useState(window.innerHeight);
+
+  useEffect(() => {
+    window.addEventListener('resize', () => {
+      setOutputWidth(getAvailableOutputWidth());
+      setOutputHeight(window.innerHeight);
+    });
+    return () =>
+      window.removeEventListener('resize', () => {
+        setOutputWidth(getAvailableOutputWidth());
+        setOutputHeight(window.innerHeight);
+      });
+  }, []);
+
   if (consoleVisible) {
     return (
       <HorizontalOutput
-        height={800}
-        width={availableWidth}
+        height={outputHeight}
+        width={outputWidth}
         setOutputHeight={() => {}}
       />
     );
